# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asset Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/menus": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get menu structure",
                "operationId": "54c2d937e1a32b13f880b49ce0444f34",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Dashboard"
                                            },
                                            "url": {
                                                "type": "string",
                                                "example": "/dashboard"
                                            },
                                            "icon": {
                                                "type": "string",
                                                "example": "dashboard-icon"
                                            },
                                            "children": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Submenu"
                                                        },
                                                        "url": {
                                                            "type": "string",
                                                            "example": "/dashboard/submenu"
                                                        },
                                                        "icon": {
                                                            "type": "string",
                                                            "example": "submenu-icon"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menus"
                ],
                "summary": "Create menu",
                "operationId": "31e1ddbd8da961fb496e3e2d46b53823",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dashboard"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "fas fa-home"
                                    },
                                    "permission_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menus/{id}": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get menu detail",
                "operationId": "f21bb63f9fe1ea437a502e5116af01a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Menus"
                ],
                "summary": "Update menu",
                "operationId": "88cfe38e1bda9aad62dbca6007b25422",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Menu"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Menus"
                ],
                "summary": "Delete menu",
                "operationId": "e367f4eb9f2883c92645c50bbac63925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu deleted"
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Menu": {
                "title": "Menu",
                "description": "Schema untuk Menu",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dashboard"
                    },
                    "url": {
                        "type": "string",
                        "example": "/dashboard"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fas fa-home"
                    },
                    "permission_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-03T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-03T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}